Initial commit
This script draws heavily on previous works. I embelished it, since I believe that some large number of websites which will decrypt your Cisco-7 strings are likely also capturing your passwords, to add them to the hackers scripts. I'd much rather have something "local" that doesn't depend on the cloud. The argument is a filename, but you can also use the wildcard (asterisk) to have it search a bunch of files at once. If there is a hit, the results show you which filename contained the hit(s). If you use the Verbose flag (-v), it will pass all of the data found in the file, EXCEPT that it will replace the ancrypted passwords with the decrypted text.